Using correct name for storing selected folder states
Renamed `repoFolder` to `selectedFolder` for storing parent folder which user selects when cloning repo.
And, renamed `updateRemoteRepoFolder` to `repoFolder` for storing folder for which user is updating remote url because in this case it is actual repoFolder unlike in case of clone repo functionality where user selects parent folder where they want their repo to be cloned.