Added --quiet flag for git clone command
This is done to avoid it logging output to stderr. So, now when running git clone command if any error occurs only it will be sent to error stream and not some stdout which is on by default if this flag is not supplied. Not sure the reason behind this mess. But, avoiding the earlier ugly hack by adding this flag.(--quiet).