Using electron built-in updater for auto-update feature
In-build autoUpdater of electron will use squirrel mac behind the scene and download the latest update for us. However, this is not supported in Linux. So, for linux we will be handling the update flow using our api and based on the response we will show prompt to users in Linux asking to download latest version. They have also option to skip the release.