Add duplicate key logic GenerateKey screen
Before generating key, we need to notify users based on their current provider and username. This is because with current implementation of checking if key already exists logic based on rsa files we might miss out on case where user has already created ssh key for it in SINGLE mode and file checking logic might miss it because of mode logic. So, we're checking in keyStore where we store both provider and username as well to confirm this case and notify user about it.