Fix : Check mode of existing username and provider combination
This is to avoid generating key for same username and provider but in different mode.
E.x. User has already created one ssh key using particular username and provider conbination in that case key would have been generated in SINGLE mode. So, when user tries to generate key for same combination considering we already have key for same provider we set the mode to MULTI thereby accidentally creating another ssh key for same combination but in MULTI mode.
So, in order to avoid such scenario we're now checking for same username and provider combination in store and if it exists we set mode to SINGLE so that user is asked to override key and thereby avoid generate key with same combination of username and provider but in different MODE.

TL;DR: We set correct mode in sessionStore based on whether keys exists with same username and provider or not.