Add connecting-account ipc event for copy link case
This ipc event needs to be added because there is some delay in case of github where we call our API to get token. So, for users who have copied key, pasted in the browser and completed the connect account process they will be redirected back to app but 'Connect' button would stay in default state while our handle app url method is getting token in exchange for code and state from our api in case of github. It might confuse them if api is really slow causing them to close the screen. So, it is better we give indicating of connecting by sending 'connecting-account' from main process moment we receive deep link url inside our handleAppUrl method.